Implement balance tracking system with connected with simple trade ca…

+++ API/Models/Bot.cs
@@ -5,9 +5,9 @@ namespace API.Models
     [Serializable]
     public class Bot
     {
-        public readonly static int DEFAULT_BALANCE = 100;
+        public readonly static double DEFAULT_BALANCE = 100;
         public DateTime Created { get; set; } = DateTime.Now;
-        public int Balance { get; set; } = DEFAULT_BALANCE;
+        public double Balance { get; set; } = DEFAULT_BALANCE;
         public required string Name { get; set; }
         public required string AuthenticationKey { get; set; }
         public bool isAdmin { get; set; } = false;
